Playback: extract candidate ranking and source search from TorrentPlaybackService
Moves the pure candidate ranking / probing surface to
PlaybackCandidateRanking and the source-search fetchers and curation
passes to PlaybackSourceSearch under lib/services/torrent_playback/.
Bodies are verbatim; see the PR's origin-diff table for the renames.

Replayed by history rewrite (approved as Option B on #238) on top of
three origin pins: the ranking pin, the source-search/engine pin and the
alias-warmup pin. The last two were authored after the first move
attempt (evidence ref evidence/t3-first-move = 3e0873a9) and placed
before the move; that order is not historical execution evidence.

To keep all three pins passing unedited across the move, the host keeps
21 @visibleForTesting constant tear-offs (18 from the first attempt plus
warmSourceAliases, searchCuratedSources and searchSeriesPackSources).
Lane T3-F repoints the suites and deletes the block.
